requirements.txt
streamlit>=1.35.0
groq>=0.9.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
plotly>=5.22.0
scikit-learn>=1.5.0
scipy>=1.13.0
requests>=2.31.0
pydantic>=2.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
